+++ src/main/java/group/msg/jpowermonitor/config/PathElement.java
@@ -3,7 +3,6 @@ package group.msg.jpowermonitor.config;
 import lombok.Data;
 import org.jetbrains.annotations.Nullable;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -13,5 +12,5 @@ import java.util.List;
 public class PathElement {
     List<String> path;
     @Nullable
-    BigDecimal energyInIdleMode;
+    Double energyInIdleMode;
 }
